\name{seerSurv-package}
\alias{seerSurv-package}
\alias{seerSurv}
\docType{package}
\title{
\packageTitle{seerSurv}
}
\description{
\packageDescription{seerSurv}
}
\details{
The main entry point is \code{\link{run_tumour_analysis}}.
See the vignette \code{vignette("seerSurv-intro", package = "seerSurv")} for a
step-by-step walkthrough.
}
\author{
\packageAuthor{seerSurv}

Maintainer: \packageMaintainer{seerSurv}
}
\references{
Mansoori S, Pandey S, Rani R, Singh B, Kurt M (2025).
"Deriving Cancer Progression Rates After Local/Regional Recurrence Using
Aggregate Survival Data: An R Package for Health Economic Evaluations."
\emph{Value in Health} (submitted).

Surveillance, Epidemiology, and End Results (SEER) Program
\url{https://seer.cancer.gov/}.
}
\keyword{package}
\seealso{
\code{\link{run_tumour_analysis}}, \code{\link{lifetable_seer}},
\code{\link{tumour_data_seer}}
}
